+++ ens-20171110/alibabacloud_ens20171110/models.py
@@ -4880,16 +4880,24 @@ class CreateClassicNetworkResponse(TeaModel):
         return self
 
 
-class CreateClusterRequest(TeaModel):
+class CreateClusterRequestControlPlaneConfig(TeaModel):
     def __init__(
         self,
-        cluster_version: str = None,
-        name: str = None,
+        image_id: str = None,
+        instance_spec: str = None,
+        node_port_range: str = None,
+        runtime: str = None,
+        size: int = None,
+        system_disk_category: str = None,
+        system_disk_size: int = None,
     ):
-        # The version of the cluster.
-        self.cluster_version = cluster_version
-        # The name of the cluster.
-        self.name = name
+        self.image_id = image_id
+        self.instance_spec = instance_spec
+        self.node_port_range = node_port_range
+        self.runtime = runtime
+        self.size = size
+        self.system_disk_category = system_disk_category
+        self.system_disk_size = system_disk_size
 
     def validate(self):
         pass
@@ -4900,18 +4908,258 @@ class CreateClusterRequest(TeaModel):
             return _map
 
         result = dict()
-        if self.cluster_version is not None:
-            result['ClusterVersion'] = self.cluster_version
-        if self.name is not None:
-            result['Name'] = self.name
+        if self.image_id is not None:
+            result['ImageId'] = self.image_id
+        if self.instance_spec is not None:
+            result['InstanceSpec'] = self.instance_spec
+        if self.node_port_range is not None:
+            result['NodePortRange'] = self.node_port_range
+        if self.runtime is not None:
+            result['Runtime'] = self.runtime
+        if self.size is not None:
+            result['Size'] = self.size
+        if self.system_disk_category is not None:
+            result['SystemDiskCategory'] = self.system_disk_category
+        if self.system_disk_size is not None:
+            result['SystemDiskSize'] = self.system_disk_size
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('ImageId') is not None:
+            self.image_id = m.get('ImageId')
+        if m.get('InstanceSpec') is not None:
+            self.instance_spec = m.get('InstanceSpec')
+        if m.get('NodePortRange') is not None:
+            self.node_port_range = m.get('NodePortRange')
+        if m.get('Runtime') is not None:
+            self.runtime = m.get('Runtime')
+        if m.get('Size') is not None:
+            self.size = m.get('Size')
+        if m.get('SystemDiskCategory') is not None:
+            self.system_disk_category = m.get('SystemDiskCategory')
+        if m.get('SystemDiskSize') is not None:
+            self.system_disk_size = m.get('SystemDiskSize')
+        return self
+
+
+class CreateClusterRequest(TeaModel):
+    def __init__(
+        self,
+        cluster_type: str = None,
+        cluster_version: str = None,
+        container_cidr: str = None,
+        control_plane_config: CreateClusterRequestControlPlaneConfig = None,
+        ens_region_id: str = None,
+        kubernetes_version: str = None,
+        load_balancer_id: str = None,
+        name: str = None,
+        pod_vswitch_ids: List[str] = None,
+        profile: str = None,
+        public_access: bool = None,
+        service_cidr: str = None,
+        vpc_id: str = None,
+        vswitch_ids: List[str] = None,
+    ):
+        # This parameter is required.
+        self.cluster_type = cluster_type
+        # The version of the cluster.
+        self.cluster_version = cluster_version
+        self.container_cidr = container_cidr
+        self.control_plane_config = control_plane_config
+        self.ens_region_id = ens_region_id
+        self.kubernetes_version = kubernetes_version
+        self.load_balancer_id = load_balancer_id
+        # The name of the cluster.
+        self.name = name
+        self.pod_vswitch_ids = pod_vswitch_ids
+        # This parameter is required.
+        self.profile = profile
+        self.public_access = public_access
+        self.service_cidr = service_cidr
+        self.vpc_id = vpc_id
+        self.vswitch_ids = vswitch_ids
+
+    def validate(self):
+        if self.control_plane_config:
+            self.control_plane_config.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.cluster_type is not None:
+            result['ClusterType'] = self.cluster_type
+        if self.cluster_version is not None:
+            result['ClusterVersion'] = self.cluster_version
+        if self.container_cidr is not None:
+            result['ContainerCidr'] = self.container_cidr
+        if self.control_plane_config is not None:
+            result['ControlPlaneConfig'] = self.control_plane_config.to_map()
+        if self.ens_region_id is not None:
+            result['EnsRegionId'] = self.ens_region_id
+        if self.kubernetes_version is not None:
+            result['KubernetesVersion'] = self.kubernetes_version
+        if self.load_balancer_id is not None:
+            result['LoadBalancerId'] = self.load_balancer_id
+        if self.name is not None:
+            result['Name'] = self.name
+        if self.pod_vswitch_ids is not None:
+            result['PodVswitchIds'] = self.pod_vswitch_ids
+        if self.profile is not None:
+            result['Profile'] = self.profile
+        if self.public_access is not None:
+            result['PublicAccess'] = self.public_access
+        if self.service_cidr is not None:
+            result['ServiceCidr'] = self.service_cidr
+        if self.vpc_id is not None:
+            result['VpcId'] = self.vpc_id
+        if self.vswitch_ids is not None:
+            result['VswitchIds'] = self.vswitch_ids
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ClusterType') is not None:
+            self.cluster_type = m.get('ClusterType')
         if m.get('ClusterVersion') is not None:
             self.cluster_version = m.get('ClusterVersion')
+        if m.get('ContainerCidr') is not None:
+            self.container_cidr = m.get('ContainerCidr')
+        if m.get('ControlPlaneConfig') is not None:
+            temp_model = CreateClusterRequestControlPlaneConfig()
+            self.control_plane_config = temp_model.from_map(m['ControlPlaneConfig'])
+        if m.get('EnsRegionId') is not None:
+            self.ens_region_id = m.get('EnsRegionId')
+        if m.get('KubernetesVersion') is not None:
+            self.kubernetes_version = m.get('KubernetesVersion')
+        if m.get('LoadBalancerId') is not None:
+            self.load_balancer_id = m.get('LoadBalancerId')
         if m.get('Name') is not None:
             self.name = m.get('Name')
+        if m.get('PodVswitchIds') is not None:
+            self.pod_vswitch_ids = m.get('PodVswitchIds')
+        if m.get('Profile') is not None:
+            self.profile = m.get('Profile')
+        if m.get('PublicAccess') is not None:
+            self.public_access = m.get('PublicAccess')
+        if m.get('ServiceCidr') is not None:
+            self.service_cidr = m.get('ServiceCidr')
+        if m.get('VpcId') is not None:
+            self.vpc_id = m.get('VpcId')
+        if m.get('VswitchIds') is not None:
+            self.vswitch_ids = m.get('VswitchIds')
+        return self
+
+
+class CreateClusterShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        cluster_type: str = None,
+        cluster_version: str = None,
+        container_cidr: str = None,
+        control_plane_config_shrink: str = None,
+        ens_region_id: str = None,
+        kubernetes_version: str = None,
+        load_balancer_id: str = None,
+        name: str = None,
+        pod_vswitch_ids_shrink: str = None,
+        profile: str = None,
+        public_access: bool = None,
+        service_cidr: str = None,
+        vpc_id: str = None,
+        vswitch_ids_shrink: str = None,
+    ):
+        # This parameter is required.
+        self.cluster_type = cluster_type
+        # The version of the cluster.
+        self.cluster_version = cluster_version
+        self.container_cidr = container_cidr
+        self.control_plane_config_shrink = control_plane_config_shrink
+        self.ens_region_id = ens_region_id
+        self.kubernetes_version = kubernetes_version
+        self.load_balancer_id = load_balancer_id
+        # The name of the cluster.
+        self.name = name
+        self.pod_vswitch_ids_shrink = pod_vswitch_ids_shrink
+        # This parameter is required.
+        self.profile = profile
+        self.public_access = public_access
+        self.service_cidr = service_cidr
+        self.vpc_id = vpc_id
+        self.vswitch_ids_shrink = vswitch_ids_shrink
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.cluster_type is not None:
+            result['ClusterType'] = self.cluster_type
+        if self.cluster_version is not None:
+            result['ClusterVersion'] = self.cluster_version
+        if self.container_cidr is not None:
+            result['ContainerCidr'] = self.container_cidr
+        if self.control_plane_config_shrink is not None:
+            result['ControlPlaneConfig'] = self.control_plane_config_shrink
+        if self.ens_region_id is not None:
+            result['EnsRegionId'] = self.ens_region_id
+        if self.kubernetes_version is not None:
+            result['KubernetesVersion'] = self.kubernetes_version
+        if self.load_balancer_id is not None:
+            result['LoadBalancerId'] = self.load_balancer_id
+        if self.name is not None:
+            result['Name'] = self.name
+        if self.pod_vswitch_ids_shrink is not None:
+            result['PodVswitchIds'] = self.pod_vswitch_ids_shrink
+        if self.profile is not None:
+            result['Profile'] = self.profile
+        if self.public_access is not None:
+            result['PublicAccess'] = self.public_access
+        if self.service_cidr is not None:
+            result['ServiceCidr'] = self.service_cidr
+        if self.vpc_id is not None:
+            result['VpcId'] = self.vpc_id
+        if self.vswitch_ids_shrink is not None:
+            result['VswitchIds'] = self.vswitch_ids_shrink
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ClusterType') is not None:
+            self.cluster_type = m.get('ClusterType')
+        if m.get('ClusterVersion') is not None:
+            self.cluster_version = m.get('ClusterVersion')
+        if m.get('ContainerCidr') is not None:
+            self.container_cidr = m.get('ContainerCidr')
+        if m.get('ControlPlaneConfig') is not None:
+            self.control_plane_config_shrink = m.get('ControlPlaneConfig')
+        if m.get('EnsRegionId') is not None:
+            self.ens_region_id = m.get('EnsRegionId')
+        if m.get('KubernetesVersion') is not None:
+            self.kubernetes_version = m.get('KubernetesVersion')
+        if m.get('LoadBalancerId') is not None:
+            self.load_balancer_id = m.get('LoadBalancerId')
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
+        if m.get('PodVswitchIds') is not None:
+            self.pod_vswitch_ids_shrink = m.get('PodVswitchIds')
+        if m.get('Profile') is not None:
+            self.profile = m.get('Profile')
+        if m.get('PublicAccess') is not None:
+            self.public_access = m.get('PublicAccess')
+        if m.get('ServiceCidr') is not None:
+            self.service_cidr = m.get('ServiceCidr')
+        if m.get('VpcId') is not None:
+            self.vpc_id = m.get('VpcId')
+        if m.get('VswitchIds') is not None:
+            self.vswitch_ids_shrink = m.get('VswitchIds')
         return self